internship-slack-bot/requirements.txt
slack_sdk>=3.27.0
slack_bolt>=1.18.0
python-dotenv>=1.0.0
pytest>=8.0.0
